+++ src/error.rs
@@ -11,33 +11,43 @@ pub struct Error {
 /// Error kinds that can occur during caching.
 #[derive(Clone, Eq, PartialEq, Debug, Fail)]
 pub enum ErrorKind {
+    /// Arises when the resource looks like a local file but it doesn't exist.
     #[fail(
         display = "Treated resource as local file, but file does not exist ({})",
         _0
     )]
     ResourceNotFound(String),
 
+    /// Arises when the resource looks like a URL, but is invalid.
+    #[fail(display = "Unable to parse resource URL ({})", _0)]
+    InvalidUrl(String),
+
+    /// Arises when the cache is being used in offline mode, but it couldn't locate
+    /// any cached versions of a remote resource.
     #[fail(
         display = "Offline mode is enabled but no cached versions of resouce exist ({})",
         _0
     )]
     NoCachedVersions(String),
 
-    #[fail(display = "Unable to parse resource URL ({})", _0)]
-    InvalidUrl(String),
-
+    /// Any IO error that could arise while attempting to cache a remote resource.
     #[fail(display = "An IO error occurred: {:?}", _0)]
     IoError(Option<io::ErrorKind>),
 
+    /// Arises when a bad HTTP status code is received while attempting to fetch
+    /// a remote resource.
     #[fail(display = "HTTP response had status code {}", _0)]
     HttpStatusError(u16),
 
+    /// Arises when an HTTP timeout error occurs while attempting to fetch a remote resource.
     #[fail(display = "HTTP response timeout out")]
     HttpTimeoutError,
 
+    /// Arises when the HTTP client fails to build.
     #[fail(display = "HTTP builder error")]
     HttpBuilderError,
 
+    /// Any other HTTP error that could occur while attempting to fetch a remote resource.
     #[fail(display = "An HTTP error occurred")]
     HttpError,
 }