+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - API is now syncronous
+- `root` configuration option renamed to `dir`.
 
 ## v0.2.0